添加评论软删除操作，修改评论和收藏接口代码

+++ app/Api/Controllers/FavoriteController.php
@@ -2,10 +2,8 @@
 
 namespace App\Api\Controllers;
 
-use App\Models\Module;
 use App\Models\Favorite;
-use App\Models\File;
-use DB;
+use App\Models\Module;
 use Exception;
 use Request;
 
@@ -14,8 +12,8 @@ class FavoriteController extends BaseController
 {
     public function transform($favorite)
     {
-        $attributes['favorite_id'] = $favorite->id;
         $attributes = $favorite->refer->getAttributes();
+        $attributes['favorite_id'] = $favorite->id;
         $attributes['images'] = $favorite->refer->images()->transform(function ($item) use ($favorite) {
             return [
                 'id' => $item->id,
@@ -24,8 +22,9 @@ class FavoriteController extends BaseController
                 'summary' => $item->summary,
             ];
         });
-        $attributes['time'] = $favorite->refer->published_at->format('m-d H:i');
-        $attributes['time_trans'] = time_trans(strtotime($favorite->refer->published_at));
+        foreach ($favorite->refer->getDates() as $date) {
+            $attributes[$date] = empty($favorite->refer->$date) ? '' : $favorite->refer->$date->toDateTimeString();
+        }
         return $attributes;
     }
 
@@ -35,6 +34,7 @@ class FavoriteController extends BaseController
      *   summary="获取收藏列表",
      *   tags={"/favorites 收藏"},
      *   @SWG\Parameter(name="site_id", in="query", required=true, description="站点ID", type="string"),
+     *   @SWG\Parameter(name="type", in="query", required=true, description="类型", type="integer"),
      *   @SWG\Parameter(name="page_size", in="query", required=true, description="分页大小", type="integer"),
      *   @SWG\Parameter(name="page", in="query", required=true, description="分页序号", type="integer"),
      *   @SWG\Parameter(name="token", in="query", required=true, description="token", type="string"),
@@ -53,6 +53,7 @@ class FavoriteController extends BaseController
         $site_id = Request::get('site_id') ? Request::get('site_id') : 1;
         $page_size = Request::get('page_size');
         $page = Request::get('page');
+        $type = Request::get('type');
 
         try {
             $member = \JWTAuth::parseToken()->authenticate();
@@ -63,7 +64,13 @@ class FavoriteController extends BaseController
             return $this->responseError('无效的token,请重新登录');
         }
 
+        $module = Module::find($type);
+        if (!$module) {
+            return $this->responseError('此类型不存在');
+        }
+
         $favorites = Favorite::where('site_id', $site_id)
+            ->where('refer_type', $module->model_class)
             ->where('member_id', $member->id)
             ->orderBy('created_at', 'desc')
             ->skip(($page - 1) * $page_size)
@@ -118,10 +125,12 @@ class FavoriteController extends BaseController
         }
 
         $module = Module::find($type);
+        if (!$module) {
+            return $this->responseError('此类型不存在');
+        }
 
         //检查收藏记录数是否过多
-        $count = DB::table('favorites')
-            ->where('refer_id', $id)
+        $count = Favorite::where('refer_id', $id)
             ->where('refer_type', $module->model_class)
             ->where('member_id', $member->id)
             ->count();
@@ -131,31 +140,29 @@ class FavoriteController extends BaseController
         }
 
         //检查总记录数是否过多
-        $count = DB::table('favorites')
-            ->where('refer_id', $id)
+        $count = Favorite::where('refer_id', $id)
             ->where('refer_type', $module->model_class)
             ->count();
         if ($count >= 1000 * 1000) {
             return $this->responseError('收藏数量过多');
         }
 
-        $__singular__ = call_user_func([$module->model_class, 'find'], $id);
+        $model = call_user_func([$module->model_class, 'find'], $id);
 
-        if (empty($__singular__)) {
+        if (empty($model)) {
             return $this->responseError('此ID不存在');
         }
 
         //判断此收藏是否已存在
         if (!Favorite::where('member_id', $member->id)->where('refer_id', $id)->where('refer_type', $module->model_class)->exists()) {
             //增加收藏数
-            $__singular__->favorites += 1;
-            $__singular__->save();
+            $model->increment('favorites');
 
             //增加收藏记录
             $favorite = new Favorite();
-            $favorite->site_id = $__singular__->site_id;
-            $favorite->refer_id = $__singular__->id;
-            $favorite->refer_type = $__singular__->getMorphClass();
+            $favorite->site_id = $model->site_id;
+            $favorite->refer_id = $model->id;
+            $favorite->refer_type = $model->getMorphClass();
             $favorite->member_id = $member->id;
 
             $favorite->save();
@@ -183,6 +190,8 @@ class FavoriteController extends BaseController
      */
     public function destroy()
     {
+        $favorite_ids = Request::get('favorite_ids');
+
         try {
             $member = \JWTAuth::parseToken()->authenticate();
             if (!$member) {
@@ -192,10 +201,9 @@ class FavoriteController extends BaseController
             return $this->responseError('无效的token,请重新登录');
         }
 
-        $favorite_ids = Request::get('favorite_ids');
-
-        DB::table('favorites')->where('member_id', $member->id)
-            ->whereIn('id', explode(',', $favorite_ids))->delete();
+        Favorite::where('member_id', $member->id)
+            ->whereIn('id', explode(',', $favorite_ids))
+            ->delete();
 
         return $this->responseSuccess();
     }
@@ -233,13 +241,16 @@ class FavoriteController extends BaseController
         }
 
         $module = Module::find($type);
+        if (!$module) {
+            return $this->responseError('此类型不存在');
+        }
 
         $favorite = Favorite::where('refer_id', $id)
             ->where('refer_type', $module->model_class)
             ->where('member_id', $member->id)
             ->first();
 
-        if($favorite){
+        if ($favorite) {
             $favorite->delete();
         }
 
@@ -279,6 +290,9 @@ class FavoriteController extends BaseController
         }
 
         $module = Module::find($type);
+        if (!$module) {
+            return $this->responseError('此类型不存在');
+        }
 
         $favorite = Favorite::where('refer_id', $id)
             ->where('refer_type', $module->model_class)
